pyproject.toml
[tool.poetry]
name = "pm64openworldrandomizer_website"
version = "0.1.0"
description = ""
authors = ["Josh <54910913+JoshW-7@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
